build: update dependencyUpdate configuration
