+++ .github/workflows/update_dependencies.yml
@@ -17,7 +17,7 @@ jobs:
       - name: Gradle dependency updates raport
         uses: eskatos/gradle-command-action@v1
         with:
-            arguments: dependencyUpdates -DoutputFormatter=json -DoutputDir=..
+            arguments: dependencyUpdates -DoutputFormatter=json -DoutputDir=.
 
       - name: Gradle Build flankScripts and add it to PATH
         run: |
@@ -35,7 +35,7 @@ jobs:
           commit-message: "[Automatic PR] Dependencies update"
           signoff: false
           branch: 'dependencies_update'
-          title: 'build: Dependencies updats'
+          title: 'build: Dependencies updates'
           body: "Dependencies updates"
           labels: |
             automated pr
