+++ build.gradle.kts
@@ -1,4 +1,4 @@
-
+import com.github.benmanes.gradle.versions.updates.DependencyUpdatesTask
 
 // Fix Exception in thread "main" java.lang.NoSuchMethodError: com.google.common.hash.Hashing.crc32c()Lcom/google/common/hash/HashFunction;
 // https://stackoverflow.com/a/45286710
@@ -12,9 +12,38 @@ configurations.all {
 
 plugins {
     kotlin(Plugins.Kotlin.PLUGIN_JVM) version Versions.KOTLIN
+    id(Plugins.BEN_MANES_PLUGIN) version Versions.BEN_MANES
 }
 
 repositories {
     jcenter()
     mavenCentral()
+}
+
+tasks.named("dependencyUpdates", DependencyUpdatesTask::class.java).configure {
+
+    gradleReleaseChannel = "release-candidate"
+
+    fun isStable(version: String) = listOf("RELEASE", "FINAL", "GA")
+            .any { version.toUpperCase().contains(it) } || "^[0-9,.v-]+(-r)?$".toRegex().matches(version)
+
+    fun isNonStable(version: String) = isStable(version).not()
+
+    resolutionStrategy {
+        componentSelection {
+            all {
+
+                if (candidate.group == "com.google.apis" &&
+                    candidate.module == "google-api-services-toolresults" &&
+                    !candidate.version.startsWith("v1beta3-")
+                ) {
+                    reject("com.google.apis:google-api-services-toolresults should use beta only")
+                }
+
+                if (isNonStable(candidate.version) && isStable(currentVersion)) {
+                    reject("Release candidate")
+                }
+            }
+        }
+    }
 }
